Refactor the SHA-2 hash family classes
