+++ CMakeLists.txt
@@ -1,7 +1,7 @@
 cmake_minimum_required(VERSION 3.28)
 project(custom_tls)
 
-set(CMAKE_CXX_STANDARD 17)
+set(CMAKE_CXX_STANDARD 20)
 set(CMAKE_CXX_STANDARD_REQUIRED ON)
 
 if (MSVC)
@@ -25,7 +25,6 @@ file(GLOB_RECURSE SOURCES
         src/mpz.cpp
         src/rsa.cpp
         src/sha1.cpp
-        src/sha2.cpp
 )
 
 file(GLOB_RECURSE TEST_SOURCES
