+++ src/sha1.cpp
@@ -2,24 +2,13 @@
 // Created by wtchr on 8/13/2024.
 //
 
-#include "tls/sha1.h"
+#include "tls/sha/sha1.h"
 
 #include <algorithm>
 #include "tls/mpz.h"
 
 static uint32_t left_rotate(const uint32_t a, const int bits) {
     return a << bits | a >> 32 - bits;
-}
-
-template<int BLOCK_SIZE>
-static void preprocess(std::vector<unsigned char> &v) {
-    const size_t len = v.size();
-    v.push_back(0x80);
-    size_t padding_size = BLOCK_SIZE - (len + 1) % BLOCK_SIZE;
-    if (padding_size < BLOCK_SIZE / 8)
-        padding_size += BLOCK_SIZE;
-    v.resize(len + 1 + padding_size, 0);
-    mpz2bnd(static_cast<unsigned long>(len * 8), v.end() - BLOCK_SIZE / 8, v.end());
 }
 
 // sha1
@@ -30,7 +19,13 @@ sha1::sha1() {
 }
 
 void sha1::preprocess(std::vector<unsigned char> &v) {
-    ::preprocess<block_size>(v);
+    const size_t len = v.size();
+    v.push_back(0x80);
+    size_t padding_size = block_size - (len + 1) % block_size;
+    if (padding_size < block_size / 8)
+        padding_size += block_size;
+    v.resize(len + 1 + padding_size, 0);
+    mpz2bnd(static_cast<unsigned long>(len * 8), v.end() - block_size / 8, v.end());
 }
 
 void sha1::process_chunk(unsigned char *p) {
